refactor: move topology config files into topologyConfigs/ subdirectory
Mirrors the llmAcceleratorConfigs/ layout ahead of the tab migration. Pure
move plus relative-import depth fixes; no behavior change.

RHOAIENG-80016
